Change content to one file
