+++ src/modules/__tests__/add-removeTask.test.js
@@ -1,6 +1,23 @@
 /** * @jest-environment jsdom */
-
+import UIDisplay from '../userInterface.js';
 import removeTask from '../removeTask.js';
+
+test('Add a new task', () => {
+  document.body.innerHTML = '<ul id="task-list"></ul>';
+
+  const newTask = {
+    description: 'desc',
+    completed: false,
+    index: 0,
+  };
+
+  UIDisplay.createTask(newTask);
+  const list = document.querySelectorAll('.task');
+
+  expect(list).toHaveLength(1);
+});
+
+
 
 test('Remove a task', () => {
   document.body.innerHTML = '<ul id="task-list"><li class="task"><input type="checkbox" class="checkB"><input class="task-d"><i class="fa-solid fa-ellipsis-vertical show"></i><i class="fa-solid fa-trash-can"></i></li></ul>';
@@ -15,5 +32,3 @@ test('Remove a task', () => {
   const list = document.querySelectorAll('.task');
   expect(list).toHaveLength(0);
 });
-
-// test ('Remove a task')
